requirements.txt
# Core dependencies
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
pydantic>=2.10.0
pydantic-settings>=2.7.0
python-dotenv>=1.0.0

# Web scraping
beautifulsoup4>=4.12.0
playwright>=1.41.0
requests>=2.31.0
httpx>=0.26.0

# Data processing
pyyaml>=6.0

# LLM integration
openai>=1.54.0

# Video generation
moviepy>=1.0.3
pydub>=0.25.1
gtts>=2.5.0  # Google Text-to-Speech (free)

# Social media APIs
instagrapi>=2.1.0  # Instagram
tweepy>=4.14.0     # Twitter/X
# tiktok-api - No official API, using manual curation in code

# Database
sqlalchemy>=2.0.0

# Email
sendgrid>=6.11.0

# Background tasks
apscheduler>=3.10.0
celery>=5.3.0
redis>=5.0.0

# Web framework
jinja2>=3.1.0
python-multipart>=0.0.6

# Utilities
loguru>=0.7.0
arrow>=1.3.0
